Move msg_server into keepers
