+++ x/cdp/keeper/msg_server.go
@@ -1,35 +1,63 @@
-package hard
+package keeper
 
 import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/kava-labs/kava/x/hard/keeper"
-	"github.com/kava-labs/kava/x/hard/types"
+	"github.com/kava-labs/kava/x/cdp/types"
 )
 
 type msgServer struct {
-	keeper keeper.Keeper
+	keeper Keeper
 }
 
 // NewMsgServerImpl returns an implementation of the swap MsgServer interface
 // for the provided Keeper.
-func NewMsgServerImpl(keeper keeper.Keeper) types.MsgServer {
+func NewMsgServerImpl(keeper Keeper) types.MsgServer {
 	return &msgServer{keeper: keeper}
 }
 
 var _ types.MsgServer = msgServer{}
 
+func (k msgServer) CreateCDP(goCtx context.Context, msg *types.MsgCreateCDP) (*types.MsgCreateCDPResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	sender, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return nil, err
+	}
+
+	err = k.keeper.AddCdp(ctx, sender, msg.Collateral, msg.Principal, msg.CollateralType)
+	if err != nil {
+		return nil, err
+	}
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			sdk.EventTypeMessage,
+			sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
+			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
+		),
+	)
+
+	id, _ := k.keeper.GetCdpID(ctx, sender, msg.CollateralType)
+	return &types.MsgCreateCDPResponse{CdpID: id}, nil
+}
+
 func (k msgServer) Deposit(goCtx context.Context, msg *types.MsgDeposit) (*types.MsgDepositResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	owner, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		return nil, err
+	}
 
 	depositor, err := sdk.AccAddressFromBech32(msg.Depositor)
 	if err != nil {
 		return nil, err
 	}
 
-	err = k.keeper.Deposit(ctx, depositor, msg.Amount)
+	err = k.keeper.DepositCollateral(ctx, owner, depositor, msg.Collateral, msg.CollateralType)
 	if err != nil {
 		return nil, err
 	}
@@ -47,12 +75,17 @@ func (k msgServer) Deposit(goCtx context.Context, msg *types.MsgDeposit) (*types
 func (k msgServer) Withdraw(goCtx context.Context, msg *types.MsgWithdraw) (*types.MsgWithdrawResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
+	owner, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		return nil, err
+	}
+
 	depositor, err := sdk.AccAddressFromBech32(msg.Depositor)
 	if err != nil {
 		return nil, err
 	}
 
-	err = k.keeper.Withdraw(ctx, depositor, msg.Amount)
+	err = k.keeper.WithdrawCollateral(ctx, owner, depositor, msg.Collateral, msg.CollateralType)
 	if err != nil {
 		return nil, err
 	}
@@ -67,30 +100,7 @@ func (k msgServer) Withdraw(goCtx context.Context, msg *types.MsgWithdraw) (*typ
 	return &types.MsgWithdrawResponse{}, nil
 }
 
-func (k msgServer) Borrow(goCtx context.Context, msg *types.MsgBorrow) (*types.MsgBorrowResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	borrower, err := sdk.AccAddressFromBech32(msg.Borrower)
-	if err != nil {
-		return nil, err
-	}
-
-	err = k.keeper.Borrow(ctx, borrower, msg.Amount)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx.EventManager().EmitEvent(
-		sdk.NewEvent(
-			sdk.EventTypeMessage,
-			sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
-			sdk.NewAttribute(sdk.AttributeKeySender, msg.Borrower),
-		),
-	)
-	return &types.MsgBorrowResponse{}, nil
-}
-
-func (k msgServer) Repay(goCtx context.Context, msg *types.MsgRepay) (*types.MsgRepayResponse, error) {
+func (k msgServer) DrawDebt(goCtx context.Context, msg *types.MsgDrawDebt) (*types.MsgDrawDebtResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
 	sender, err := sdk.AccAddressFromBech32(msg.Sender)
@@ -98,12 +108,7 @@ func (k msgServer) Repay(goCtx context.Context, msg *types.MsgRepay) (*types.Msg
 		return nil, err
 	}
 
-	owner, err := sdk.AccAddressFromBech32(msg.Owner)
-	if err != nil {
-		return nil, err
-	}
-
-	err = k.keeper.Repay(ctx, sender, owner, msg.Amount)
+	err = k.keeper.AddPrincipal(ctx, sender, msg.CollateralType, msg.Principal)
 	if err != nil {
 		return nil, err
 	}
@@ -115,7 +120,30 @@ func (k msgServer) Repay(goCtx context.Context, msg *types.MsgRepay) (*types.Msg
 			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
 		),
 	)
-	return &types.MsgRepayResponse{}, nil
+	return &types.MsgDrawDebtResponse{}, nil
+}
+
+func (k msgServer) RepayDebt(goCtx context.Context, msg *types.MsgRepayDebt) (*types.MsgRepayDebtResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	sender, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return nil, err
+	}
+
+	err = k.keeper.RepayPrincipal(ctx, sender, msg.CollateralType, msg.Payment)
+	if err != nil {
+		return nil, err
+	}
+
+	ctx.EventManager().EmitEvent(
+		sdk.NewEvent(
+			sdk.EventTypeMessage,
+			sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
+			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
+		),
+	)
+	return &types.MsgRepayDebtResponse{}, nil
 }
 
 func (k msgServer) Liquidate(goCtx context.Context, msg *types.MsgLiquidate) (*types.MsgLiquidateResponse, error) {
@@ -131,7 +159,7 @@ func (k msgServer) Liquidate(goCtx context.Context, msg *types.MsgLiquidate) (*t
 		return nil, err
 	}
 
-	err = k.keeper.AttemptKeeperLiquidation(ctx, keeper, borrower)
+	err = k.keeper.AttemptKeeperLiquidation(ctx, keeper, borrower, msg.CollateralType)
 	if err != nil {
 		return nil, err
 	}
