+++ x/hard/keeper/msg_server.go
@@ -1,64 +1,34 @@
-package cdp
+package keeper
 
 import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/kava-labs/kava/x/cdp/keeper"
-	"github.com/kava-labs/kava/x/cdp/types"
+
+	"github.com/kava-labs/kava/x/hard/types"
 )
 
 type msgServer struct {
-	keeper keeper.Keeper
+	keeper Keeper
 }
 
 // NewMsgServerImpl returns an implementation of the swap MsgServer interface
 // for the provided Keeper.
-func NewMsgServerImpl(keeper keeper.Keeper) types.MsgServer {
+func NewMsgServerImpl(keeper Keeper) types.MsgServer {
 	return &msgServer{keeper: keeper}
 }
 
 var _ types.MsgServer = msgServer{}
 
-func (k msgServer) CreateCDP(goCtx context.Context, msg *types.MsgCreateCDP) (*types.MsgCreateCDPResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	sender, err := sdk.AccAddressFromBech32(msg.Sender)
-	if err != nil {
-		return nil, err
-	}
-
-	err = k.keeper.AddCdp(ctx, sender, msg.Collateral, msg.Principal, msg.CollateralType)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx.EventManager().EmitEvent(
-		sdk.NewEvent(
-			sdk.EventTypeMessage,
-			sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
-			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
-		),
-	)
-
-	id, _ := k.keeper.GetCdpID(ctx, sender, msg.CollateralType)
-	return &types.MsgCreateCDPResponse{CdpID: id}, nil
-}
-
 func (k msgServer) Deposit(goCtx context.Context, msg *types.MsgDeposit) (*types.MsgDepositResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	owner, err := sdk.AccAddressFromBech32(msg.Owner)
-	if err != nil {
-		return nil, err
-	}
 
 	depositor, err := sdk.AccAddressFromBech32(msg.Depositor)
 	if err != nil {
 		return nil, err
 	}
 
-	err = k.keeper.DepositCollateral(ctx, owner, depositor, msg.Collateral, msg.CollateralType)
+	err = k.keeper.Deposit(ctx, depositor, msg.Amount)
 	if err != nil {
 		return nil, err
 	}
@@ -76,17 +46,12 @@ func (k msgServer) Deposit(goCtx context.Context, msg *types.MsgDeposit) (*types
 func (k msgServer) Withdraw(goCtx context.Context, msg *types.MsgWithdraw) (*types.MsgWithdrawResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
-	owner, err := sdk.AccAddressFromBech32(msg.Owner)
-	if err != nil {
-		return nil, err
-	}
-
 	depositor, err := sdk.AccAddressFromBech32(msg.Depositor)
 	if err != nil {
 		return nil, err
 	}
 
-	err = k.keeper.WithdrawCollateral(ctx, owner, depositor, msg.Collateral, msg.CollateralType)
+	err = k.keeper.Withdraw(ctx, depositor, msg.Amount)
 	if err != nil {
 		return nil, err
 	}
@@ -101,15 +66,15 @@ func (k msgServer) Withdraw(goCtx context.Context, msg *types.MsgWithdraw) (*typ
 	return &types.MsgWithdrawResponse{}, nil
 }
 
-func (k msgServer) DrawDebt(goCtx context.Context, msg *types.MsgDrawDebt) (*types.MsgDrawDebtResponse, error) {
+func (k msgServer) Borrow(goCtx context.Context, msg *types.MsgBorrow) (*types.MsgBorrowResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
-	sender, err := sdk.AccAddressFromBech32(msg.Sender)
+	borrower, err := sdk.AccAddressFromBech32(msg.Borrower)
 	if err != nil {
 		return nil, err
 	}
 
-	err = k.keeper.AddPrincipal(ctx, sender, msg.CollateralType, msg.Principal)
+	err = k.keeper.Borrow(ctx, borrower, msg.Amount)
 	if err != nil {
 		return nil, err
 	}
@@ -118,13 +83,13 @@ func (k msgServer) DrawDebt(goCtx context.Context, msg *types.MsgDrawDebt) (*typ
 		sdk.NewEvent(
 			sdk.EventTypeMessage,
 			sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
-			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
+			sdk.NewAttribute(sdk.AttributeKeySender, msg.Borrower),
 		),
 	)
-	return &types.MsgDrawDebtResponse{}, nil
+	return &types.MsgBorrowResponse{}, nil
 }
 
-func (k msgServer) RepayDebt(goCtx context.Context, msg *types.MsgRepayDebt) (*types.MsgRepayDebtResponse, error) {
+func (k msgServer) Repay(goCtx context.Context, msg *types.MsgRepay) (*types.MsgRepayResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
 	sender, err := sdk.AccAddressFromBech32(msg.Sender)
@@ -132,7 +97,12 @@ func (k msgServer) RepayDebt(goCtx context.Context, msg *types.MsgRepayDebt) (*t
 		return nil, err
 	}
 
-	err = k.keeper.RepayPrincipal(ctx, sender, msg.CollateralType, msg.Payment)
+	owner, err := sdk.AccAddressFromBech32(msg.Owner)
+	if err != nil {
+		return nil, err
+	}
+
+	err = k.keeper.Repay(ctx, sender, owner, msg.Amount)
 	if err != nil {
 		return nil, err
 	}
@@ -144,7 +114,7 @@ func (k msgServer) RepayDebt(goCtx context.Context, msg *types.MsgRepayDebt) (*t
 			sdk.NewAttribute(sdk.AttributeKeySender, msg.Sender),
 		),
 	)
-	return &types.MsgRepayDebtResponse{}, nil
+	return &types.MsgRepayResponse{}, nil
 }
 
 func (k msgServer) Liquidate(goCtx context.Context, msg *types.MsgLiquidate) (*types.MsgLiquidateResponse, error) {
@@ -160,7 +130,7 @@ func (k msgServer) Liquidate(goCtx context.Context, msg *types.MsgLiquidate) (*t
 		return nil, err
 	}
 
-	err = k.keeper.AttemptKeeperLiquidation(ctx, keeper, borrower, msg.CollateralType)
+	err = k.keeper.AttemptKeeperLiquidation(ctx, keeper, borrower)
 	if err != nil {
 		return nil, err
 	}
